Completly renewed the Collision system by replacing it with the Collider class - Added the collider to the player and adjusted some more settings

+++ src/main/java/fr/iut/hev/root/model/Gravity.java
@@ -1,9 +1,9 @@
 package fr.iut.hev.root.model;
 
 public class Gravity {
-    private static final double gravityForce = 1.0;
+    private static final int gravityForce = 1;
 
-    public static double getGravityForce() {
+    public static int getGravityForce() {
         return gravityForce;
     }
 }
